SAT-581: Use /etc/kubernetes/admin.conf as default kube config
Use the file `/etc/kubernetes/admin.conf` as the default value for the
kubernetes config file that is bind-mounted into the sat container under
podman. The other file `~/.kube/config` does not exist on a 1.4 system
anymore.

The `kubectl` command still expects the config to be present at
`$HOME/.kube/config`, so bind mount it into that location. It works
without being in that location on ncn-w001 because the environment
variable `KUBECONFIG` is set.

Test Description:
Tested this change on groot-ncn-w001 and verified that I could run the
following in the sat container started with `sat bash`:

* `sat k8s`
* kubectl get pods -A