SAT-687: Update default SAT_IMAGE and allow setting just tag
Update the default value of `SAT_IMAGE` and allow setting just the tag
or just the repository individually.

Also use a `sed` command in the RPM spec file to allow us to only
hard-code the default values for the repository and tag in one location
rather than in both the `sat-manpage.sh` and `sat-podman.sh` scripts.

Test Description:
Downloaded the built RPM from artifactory, installed on groot-ncn-w001,
removed any downloaded cray-sat images that were previously downloaded
with podman, ran `sat bash`, and verified that the correct image was
downloaded and run under podman.

Test the following scenarios with `sat` and `sat-man` to ensure variable
substitutions are correct:

* SAT_IMAGE unset, SAT_REPOSITORY unset, SAT_TAG unset
* SAT_IMAGE unset, SAT_REPOSITORY unset, SAT_TAG set
* SAT_IMAGE unset, SAT_REPOSITORY set, SAT_TAG unset
* SAT_IMAGE unset, SAT_REPOSITORY set, SAT_TAG set
* SAT_IMAGE set, SAT_REPOSITORY unset, SAT_TAG unset
* SAT_IMAGE set, SAT_REPOSITORY set, SAT_TAG set