SAT-662: Mount config directory in sat-podman container
This commit adds a mount between $HOME/.config/sat/ on the host
and the default configuration directory within the container.
The default configuration directory is used within the container
so that $SAT_CONFIG_FILE does not need to be set within the
container, although it is still possible to use with `sat bash`.

It also adds logic to sat-podman.sh to try to create the
configuration directory if it does not exist, and if it cannot
be created, print a warning message.

Test Description:
* Installed this version of the script along with a version
  of the `sat` docker container that automatically generates
  the configuration file upon first run of sat.  Ran `sat auth`
  using the wrapper, and verified that the configuration
  files persisted on the host filesystem.
* Next, ran other `sat` commands using the wrapper to verify
  that the configuration persisted by verifying that the commands
  worked.
* Finally, set SAT_CONFIG_DIR to the location of a file, so that a directory
  could not be created at the given path.  Verified that a warning
  was printed.