SAT-712: Add changelog and contribution guidelines
Add a changelog file called `CHANGELOG.md`, and add instructions on
contributing to this repository that include instructions on updating
the changelog with each change. The `CONTRIBUTING.md` file was copied
from the SAT/sat repo, and the line about docker version updates was
removed as it does not apply here.

Add two versions to `CHANGELOG.md`, one for everything that is currently
in master, and another for everything else.

Add a script copied from the SAT/sat repo that parses the `CHANGELOG.md`
file and extracts the latest version number from it, and use that script
in the spec file to get the version, so it only has to be specified in
the one place in the changelog.

Test Description:
Pushed branch, checked RPM that was built to ensure it got the version
correctly. Inspected the markdown files in Bitbucket to ensure proper
formatting.