SAT-885: Add an entry to the changelog and bump version
Add an entry to the changelog and bump the version since we did change
the `DEFAULT_SAT_TAG` that gets substituted into place in the wrapper
scripts. We probably want to differentiate that this is a different
version of sat-podman because of that.

Test Description:
Ran `changelog.py ./CHANGELOG.md` and verified the correct version was
printed.