SAT-981: Create /opt/cray/etc/ when it does not exist
This commit addresses the case when ``/opt/cray/etc/`` does
not exist by attempting to create it. If it cannot be created,
then a warning is logged. After it is created, this directory
is mounted in the SAT container.

This commit also defines a version 1.4.6 in the change log.

Test Description:
* I built the cray-sat-podman RPM and installed it on an NCN
* I removed the ``/opt/cray/etc/`` directory
* I ran ``sat showrev`` to verify that the directory could be
  created and mounted.

This commit is cherry-picked from a commit to integration:
94d53ac5a5570c1154960a0fc54acd3849ba9b65