SAT-902: Use newer version of dst-shared library
Use the `release/sat_2.1.12` (underscore is actually a dash) tag of the
dst-shared library in order to get the new functionality that uploads
RPMs to arti in addition to car.  This will allow us to update the
product stream to pull RPMs from arti.

Test Description:
Pushed and watched for build to succeed.