CRAYSAT-940: Set the PAGER environment variable to "less"
This change causes the `sat-man` command to use "less" as its pager
instead of "more". This is mostly a quality-of-life improvement as more
is missing many "modern" features of less like scrolling backward etc.

Test Description:
Build RPM, convert to cpio, extract to a local directory, and run the
sat-man script to make sure there aren't issues.