CRAYSAT-1316: Fix replacement of DEFAULT_SAT_TAG in spec file
The old line for replacing the value of `DEFAULT_SAT_TAG` was left in
the spec file along with the new line. The old line should be removed so
that `DEFAULT_SAT_TAG` gets correctly replaced with the newer version.

Test Description:
Inspected built RPM and ensured that the correct cray/cray-sat container
version was specified.