CRAYSAT-1374: Mount current working directory as /sat/share
This commit changes how the current working directory is mounted in
the cray-sat container. While the current working directory was
previously mounted in the container on the same path as on the host,
it is now mounted on `/sat/share`. This prevents the mount from
accidentally mounting over container system files when the working
directory is `/` or `/usr/`.

As a part of this, a check in sat-podman.sh on $PWD being equal to
$HOME is removed, as this check to prevent duplicate mounts is no
longer needed.

Document this behavior in the `sat-podman.8` man page, explaining
how the mount works and documenting that paths now need to be relative
paths to the current working directory.

Test Description:
* I built the RPM using the Jenkins pipeline and installed it on
  mug-ncn-m001 .
* I ran `sat bash` and inspected `$HOME` and `/sat/share/` to verify
  the contents were the expected contents.
* I ran `sat init -o ./my.toml` to check writing to path relative to
  the current working directory.
* I ran `sat status` with `/root/` as my current working directory
  to make sure that the existing mount of $HOME/.config/sat/ was
  working.
* Ran `man sat-podman` on the NCN to inspect the man page changes.
* Repeated the above with the following as my working directory:
  * `/`
  * `/usr`
  * `/root/.config/sat/`