CRAYSAT-1451: Publish to sat-rpms
This commit modifies the calls to publishCsmRpms so that
the RPMs are published to the sat-rpms repository rather than the
csm-rpms repository (the default).

Also, this commit changes the container image in which the RPM
is built, as this was moved.

Test Description: Verified RPMs were published in Jenkins

(cherry picked from commit 9c6dacc68f66d9596811b1ea2a41a53251fd9665)