CRAYSAT-1444: Update license text
This commit updates all the source files in this repository to use
the MIT license and the copyright line format expected by the license
checker.

This commit also updates the value of DEFAULT_SAT_TAG to correspond
to the latest 3.14 build of the SAT container image.

Add change log entry and increment version.

Test Description: Built the RPM