CRAYSAT-1774: Build as noarch, noos, and publish to csm-rpms
This RPM delivers only shell scripts and man pages, which means it is
architecture-independent, so use noarch as the build architecture. It is
also not very closely tied with a single SLES version, so publish it as
a "noos" RPM like CSM has done for other RPMs which are not tied to
specific SLES versions. Change the Docker image used to build to the
version of SLES that matches the version on management nodes in CSM 1.5,
which is SLES 15 SP5, not that it should matter much, given the
independence from SLES versions discussed above.

In addition, publish to csm-rpms instead of to sat-rpms in preparation
for merging the SAT product into the CSM product. This will make the
RPMs available in the same repositories already in use by CSM. The `csm`
repository will be updated to start including this `cray-sat-podman` RPM
from this new location. This enables merging the SAT product into the
CSM product.

While not required, it will also be a good idea to update the
metal-provision repository to remove the old sat-rpms location once this
is merged.

Test Description:
Pushed branch and checked where Jenkins published the RPM. Downloaded
and inspected the RPM for correctness.