CRAYSAT-1767: Always use CSM path for cray-sat image
Now that the CSM product includes the Ansible content that populates the
file `/opt/cray/etc/sat/version`, we should now use the path of that
image in Nexus as it is uploaded by CSM.

In addition, the cray-sat container image has moved from sat-docker to
csm-docker in Artifactory and Nexus, so update the default image name
used by the wrapper scripts accordingly.

Note that this is a backwards incompatible change, in that the wrapper
scripts will no longer work with the versions of the cray-sat container
image provided by the SAT product. This requires CSM 1.6.

Test Description:
Tested on mug as follows. First, used `podman push` to push a stable `cray-sat:3.25.5`
image to the new csm-docker location in Nexus both as the 3.25.5 tag and
as a csm-latest tag. This simulates the expected state of Nexus after a
CSM 1.6 install. Next, downloaded the `cray-sat-podman` RPM built from
this branch, extracted it, and executed the `sat` and `sat-man` scripts
it contained under the following conditions:

* No `/opt/cray/etc/sat/version` file present. In this case, it pulls
  the csm-latest tag from the new location in Nexus.
* The value `3.25.5` written to the file `/opt/cray/etc/sat/version`. In
  this case, it pulls that tag from the new location in Nexus.

The local `podman` image storage was cleared with `podman rmi` between
testing `sat` and `sat-podman` to be sure `sat-man` was pulling and using
the image from the new location.